Set default fp32 accumulation strategy to "simple" (#438)
* compiler: default fp32 accumulation to simple

* tests: update golden references

---------

+++ tests/golden/op_negativeloglikelihoodloss_negative_log_likelihood_loss.c
@@ -4,7 +4,7 @@
  * Codegen settings:
  *   emit_testbench: False
  *   restrict_arrays: True
- *   fp32_accumulation_strategy: fp64
+ *   fp32_accumulation_strategy: simple
  *   fp16_accumulation_strategy: fp32
  *   large_temp_threshold: 1024
  *   large_weight_threshold: 102400
@@ -54,8 +54,8 @@ static inline void node0_negativeloglikelihoodloss(const float input0[2][3], con
     const idx_t n = 2;
     const idx_t c = 3;
     const idx_t d = 1;
-    double loss_sum = 0.0;
-    double weight_sum = 0.0;
+    float loss_sum = 0.0f;
+    float weight_sum = 0.0f;
     for (idx_t n_idx = 0; n_idx < n; ++n_idx) {
         for (idx_t d_idx = 0; d_idx < d; ++d_idx) {
             idx_t target_index = n_idx * d + d_idx;
@@ -64,13 +64,13 @@ static inline void node0_negativeloglikelihoodloss(const float input0[2][3], con
             } else {
                 idx_t class_index = (idx_t)target_value;
                 idx_t input_index = (n_idx * c + class_index) * d + d_idx;
-                double value = -(double)input_flat[input_index];
+                float value = -(float)input_flat[input_index];
                 loss_sum += value;
-                weight_sum += 1.0;
+                weight_sum += 1.0f;
             }
         }
     }
-    if (weight_sum == 0.0) {
+    if (weight_sum == 0.0f) {
         output_flat[0] = 0.0f;
     } else {
         output_flat[0] = loss_sum / weight_sum;
